backend/requirements.txt
fastapi
uvicorn[standard]

# Linting/formatting
black
ruff

# parser
fastapi
uvicorn
pydantic
python-multipart
pymupdf
python-docx
dateparser
python-dateutil